feat(Global): :sparkles: Refactoring & New Implementations Completed

+++ src/player/interceptors/logout.interceptor.ts
@@ -26,7 +26,7 @@ export class LogoutInterceptor implements NestInterceptor {
     const cookies: Record<string, any> = request.cookies;
     const sessionToken: unknown = cookies[COOKIE_NAMES.SESSION_TOKEN];
 
-    if (typeof sessionToken === 'string') {
+    if (typeof sessionToken === 'string' && !!sessionToken) {
       await this.tokenService.deleteToken(sessionToken);
     }
